refactor: Add barrel files for streamlined imports
- [ ] Implemented index.ts (barrel files) in component directories with multiple files.
- [ ] This simplifies import statements in other parts of the application and improves code organization.
