+++ src/RTK/index.ts
@@ -1,0 +1,2 @@
+export { store } from './store';
+export * from './slices';
