+++ src/RTK/slices/index.ts
@@ -1,0 +1,8 @@
+export {
+  cartSlice,
+  addToCart,
+  removeFromCart,
+  increaseQuantity,
+  decreaseQuantity,
+} from './cartSlice';
+export { uiSlice, openCart, closeCart, toggleCart } from './uiSlice';
