+++ src/components/cart/index.ts
@@ -1,0 +1,2 @@
+export { CartProductCard } from './CartProductCard';
+export { CartSidebar } from './CartSidebar';
