+++ src/constants/index.ts
@@ -1,0 +1,4 @@
+export { ASSISTANCE_ITEMS, SERVICE_ITEMS, SOCIAL_LINKS } from './footer';
+export { SUB_NAV_ITEMS, MAIN_NAV_LINKS, USER_ACTIONS_LINKS } from './navigation';
+export { PROMO_ITEMS } from './promotions';
+export type { PromoItemType } from './promotions';
