+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export type { CartItemType } from './cart';
+export type { ProductType, ProductVariant } from './product';
